Add conversion subpackage
