+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/converter/DefaultArgumentConverter.java
@@ -23,8 +23,8 @@ import java.util.UUID;
 
 import org.apiguardian.api.API;
 import org.junit.jupiter.api.extension.ParameterContext;
-import org.junit.platform.commons.support.ConversionException;
-import org.junit.platform.commons.support.StringConversionSupport;
+import org.junit.platform.commons.support.conversion.ConversionException;
+import org.junit.platform.commons.support.conversion.StringConversionSupport;
 import org.junit.platform.commons.util.ClassLoaderUtils;
 import org.junit.platform.commons.util.ReflectionUtils;
 
@@ -44,7 +44,7 @@ import org.junit.platform.commons.util.ReflectionUtils;
  *
  * @since 5.0
  * @see org.junit.jupiter.params.converter.ArgumentConverter
- * @see org.junit.platform.commons.support.StringConversionSupport
+ * @see org.junit.platform.commons.support.conversion.StringConversionSupport
  */
 @API(status = INTERNAL, since = "5.0")
 public class DefaultArgumentConverter implements ArgumentConverter {
