+++ junit-platform-commons/src/main/java/org/junit/platform/commons/support/conversion/ConversionException.java
@@ -8,7 +8,7 @@
  * https://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.platform.commons.support;
+package org.junit.platform.commons.support.conversion;
 
 import static org.apiguardian.api.API.Status.EXPERIMENTAL;
