+++ junit-platform-commons/src/main/java/org/junit/platform/commons/support/conversion/StringToBooleanConverter.java
@@ -8,21 +8,23 @@
  * https://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.platform.commons.support;
+package org.junit.platform.commons.support.conversion;
 
 import org.junit.platform.commons.util.Preconditions;
 
-class StringToCharacterConverter implements StringToObjectConverter {
+class StringToBooleanConverter implements StringToObjectConverter {
 
 	@Override
 	public boolean canConvert(Class<?> targetType) {
-		return targetType == Character.class;
+		return targetType == Boolean.class;
 	}
 
 	@Override
 	public Object convert(String source, Class<?> targetType) {
-		Preconditions.condition(source.length() == 1, () -> "String must have length of 1: " + source);
-		return source.charAt(0);
+		boolean isTrue = "true".equalsIgnoreCase(source);
+		Preconditions.condition(isTrue || "false".equalsIgnoreCase(source),
+			() -> "String must be 'true' or 'false' (ignoring case): " + source);
+		return isTrue;
 	}
 
 }
