+++ junit-platform-commons/src/main/java/org/junit/platform/commons/support/conversion/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Maintained conversion APIs provided by the JUnit Platform.
+ */
+
+package org.junit.platform.commons.support.conversion;
