+++ junit-platform-commons/src/module/org.junit.platform.commons/module-info.java
@@ -36,6 +36,7 @@ module org.junit.platform.commons {
 			org.junit.platform.testkit,
 			org.junit.vintage.engine;
 	exports org.junit.platform.commons.support;
+	exports org.junit.platform.commons.support.conversion;
 	exports org.junit.platform.commons.util to
 			org.junit.jupiter.api,
 			org.junit.jupiter.engine,
